Move cmd/ to core/, preparing for cobra
