+++ cmd/completion.go
