+++ core/handler.go
@@ -6,11 +6,11 @@ import (
 	"net"
 	"time"
 
-	"github.com/netm4ul/netm4ul/cmd/api"
-	"github.com/netm4ul/netm4ul/cmd/client"
 	"github.com/netm4ul/netm4ul/cmd/colors"
-	"github.com/netm4ul/netm4ul/cmd/config"
-	"github.com/netm4ul/netm4ul/cmd/server"
+	"github.com/netm4ul/netm4ul/core/api"
+	"github.com/netm4ul/netm4ul/core/client"
+	"github.com/netm4ul/netm4ul/core/config"
+	"github.com/netm4ul/netm4ul/core/server"
 )
 
 const (
