+++ core/server/server.go
@@ -14,9 +14,9 @@ import (
 	mgo "gopkg.in/mgo.v2"
 
 	"github.com/netm4ul/netm4ul/cmd/colors"
-	"github.com/netm4ul/netm4ul/cmd/config"
-	"github.com/netm4ul/netm4ul/cmd/server/database"
-	"github.com/netm4ul/netm4ul/cmd/session"
+	"github.com/netm4ul/netm4ul/core/config"
+	"github.com/netm4ul/netm4ul/core/server/database"
+	"github.com/netm4ul/netm4ul/core/session"
 )
 
 var (
